feat: dynamic browser support - remove browser http client - add browser support to node http client - naming/structure
