+++ src/function/config.ts
@@ -59,10 +59,6 @@ export class Configuration {
         return this.config.suffix;
     }
 
-    getProviderName(): string {
-        return this.config.provider;
-    }
-
     getLanguage(): 'ts' | 'js' | 'onlyJs' {
         return this.config.language
     }
