+++ src/function/copy-resources.ts
@@ -5,7 +5,8 @@ import {Configuration} from "./config";
 export const copyResources = (configuration: Configuration) => {
     const folder = configuration.getFolderManager();
     const iAjaxInterface = nodePath.join(__dirname, '..', 'static', 'interface', 'i-$-open-api.ts');
-    const httpFunction = nodePath.join(__dirname, '..', 'static', 'function', `http-${configuration.getProviderName()}.ts`);
+    const httpFunction = nodePath.join(__dirname, '..', 'static', 'function', `http.ts`);
+    const fetchNodeFunction = nodePath.join(__dirname, '..', 'static', 'function', `fetch-node.ts`);
     const queryFunction = nodePath.join(__dirname, '..', 'static', 'function', 'get-query-params.ts');
     const openApiFunction = nodePath.join(__dirname, '..', 'static', 'function', 'open-api.ts');
 
@@ -13,4 +14,5 @@ export const copyResources = (configuration: Configuration) => {
     copyFileSync(queryFunction, nodePath.join(folder.getFunctionFolder(), 'get-query-params.ts'));
     copyFileSync(openApiFunction, nodePath.join(folder.getFunctionFolder(), 'open-api.ts'));
     copyFileSync(iAjaxInterface, nodePath.join(folder.getInterfaceFolder(), 'i-$-open-api.ts'));
+    copyFileSync(fetchNodeFunction, nodePath.join(folder.getFunctionFolder(), 'fetch-node.ts'))
 }
